Adds a first take at an independent `prefect-cloud login` and `logout`
With the goal of keeping `prefect-cloud` as lightweight as possible,
we are not requiring `prefect`, `FastAPI`, or `uvicorn` to run the
Cloud login flow.  Instead, we use mostly the standard library, `httpx`,
and `toml` to implement login, API key handoff, and profile management.

This is fairly barebones in terms of TUI niceties, but it does bring
over the `prompt_select_from_list` from `prefect`.  Apart from that
one question, this attempts to follow the Unix philosophy of only
saying what's absolutely necessary.

Tests will come after we have the testing framework and suites up.

This may require PrefectHQ/nebula-ui#5539 to be merged to work correctly
with all browsers.