Uses a pull step to install Python dependencies (#57)
If we use the `pip_packages` job variable, that will end up being passed
as `EXTRA_PIP_PACKAGES`, which are installed much earlier in the Docker
entrypoint rather than in Python after we have logging established.  By
moving these to a pull step, we will be able to get better visibility
into failures with the installation.

Part of ENG-1050