Infers and autocompletes the `--from` repo when possible (#59)
Most folks will likely either be working outside of a git repo, or
within a git repo with an origin at Github.  In these cases, we can just
assume/default to the fact that they likely want _this_ repo to be their
code storage.  Here we add autocompletion and defaulting to guess the
repo when it's not provided.