Rewords the summary from `prefect-cloud deploy` to educate more
One thing we were concerned about with inferring the repo instead of
requiring `--from` is that it may not be as clear to a user that the
code must come from a remote repo.  We don't want to give them the
impression that it's their local code that's going to be deployed.

I realized that we were relying on `--from` to do a key point of
education; by making it required, we were educating the user that this
is required for MEX deployments.  `--from` was kinda noisy and
repetitive when you're already doing the standard thing of using a git
repo for your code, so I wanted to eliminate it.

Here I'm restructuring the output from `prefect-cloud deploy` to tell
the user explicitly what we're doing on their behalf (deploying a
function from a file from a repo to Prefect Cloud), and then making the
output a little simpler.